Fix some clang compilation bugs

+++ llceflib/llrenderhandler.cpp
@@ -105,7 +105,7 @@ void LLRenderHandler::OnPaint(CefRefPtr<CefBrowser> browser, PaintElementType ty
 
 void LLRenderHandler::OnCursorChange(CefRefPtr<CefBrowser> browser, CefCursorHandle cursor, CursorType type, const CefCursorInfo& custom_cursor_info)
 {
-    mParent->onCursorChanged((LLCEFLib::ECursorType)type, (unsigned int)cursor);
+    mParent->onCursorChanged((LLCEFLib::ECursorType)type, (unsigned int)(size_t)cursor);
 }
 
 void LLRenderHandler::OnPopupShow(CefRefPtr<CefBrowser> browser, bool show)
